Update OrderService to check for existing address
Changes have been made to OrderService to check if an address already exists before saving a new one. In the process, the checking and retrieval of existing addresses has also been added to AddressDao. This will prevent the creation of duplicate addresses in the database and optimize the efficiency of address handling in orders.